Moved WorkspacePath to blueprint v0.7.0 (#2089)
