+++ tests/integration/source_code/test_jobs.py
@@ -6,6 +6,7 @@ from io import StringIO
 from pathlib import Path
 from unittest.mock import create_autospec
 
+from databricks.labs.blueprint.paths import WorkspacePath
 from databricks.labs.blueprint.tui import Prompts
 from databricks.sdk.errors import NotFound
 from databricks.sdk.retries import retried
@@ -20,7 +21,6 @@ from databricks.labs.ucx.source_code.linters.files import LocalCodeLinter
 from databricks.labs.ucx.source_code.linters.context import LinterContext
 from databricks.labs.ucx.source_code.path_lookup import PathLookup
 from databricks.sdk.service import jobs, compute, pipelines
-from databricks.labs.ucx.mixins.wspath import WorkspacePath
 
 
 @retried(on=[NotFound], timeout=timedelta(minutes=2))
